# apps/server/openapi.yaml
openapi: 3.0.3
info:
  title: OpenBTS API
  description: API for OpenBTS Backend
  version: 1.0.0
servers:
  - url: https://openbts.sakilabs.com/api/v1
    description: API v1 base path
...
paths:
...
  /submissions:
    get:
      tags: [Submissions]
      summary: List submissions
      security:
        - ApiKeyAuth: []
        - CookieAuth: []
      parameters:
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 100, default: 10 }
        - name: offset
          in: query
          schema: { type: integer, minimum: 0, default: 0 }
        - name: status
          in: query
          schema: { type: string, enum: [pending, approved, rejected] }
        - name: type
          in: query
          schema: { type: string, enum: [new, update, delete] }
        - name: station_id
          in: query
          description: Filter by station's station_id identifier
          schema: { type: string }
        - name: submitter_id
          in: query
          description: Filter by submitter user ID (admin only)
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [data, totalCount]
                properties:
                  data: { type: array, items: { $ref: "#/components/schemas/SubmissionWithRelations" } }
                  totalCount: { type: integer }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "429": { $ref: "#/components/responses/TooManyRequests" }
        "500": { $ref: "#/components/responses/InternalServerError" }
    post:
      tags: [Submissions]
      summary: Create submission(s) (single object or array of 1-10)
      description: In each item's `cells[].details`, do not send **cid_long** (UMTS), **ecid** (LTE) or **nci** (NR); they are computed on the server. If any of the items in the array is invalid or will throw an error, the entire request will be rejected with an error and no submissions will be created.
      security:
        - ApiKeyAuth: []
        - CookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              oneOf:
                - $ref: "#/components/schemas/SubmissionCreateSingle"
                - type: array
                  minItems: 1
                  maxItems: 10
                  items: { $ref: "#/components/schemas/SubmissionCreateSingle" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data: { type: array, items: { $ref: "#/components/schemas/SubmissionCreateResponse" } }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "429": { $ref: "#/components/responses/TooManyRequests" }
        "500": { $ref: "#/components/responses/InternalServerError" }
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-Api-Key
      description: API Key for authentication via X-Api-Key header
    CookieAuth:
      type: apiKey
      in: cookie
      name: auth
      description: Cookie-based authentication
...
  responses:
    BadRequest:
      description: Bad request (invalid parameters, validation error, or invalid query)
      content:
        application/json:
          schema:
            allOf:
              - $ref: "#/components/schemas/ApiError"
              - type: object
                properties:
                  errors:
                    items:
                      properties:
                        code:
                          enum: [BAD_REQUEST, VALIDATION_ERROR, INVALID_QUERY]
          example:
            errors:
              - code: BAD_REQUEST
                message: Invalid request
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            allOf:
              - $ref: "#/components/schemas/ApiError"
              - type: object
                properties:
                  errors:
                    items:
                      properties:
                        code:
                          enum: [UNAUTHORIZED]
          example:
            errors:
              - code: UNAUTHORIZED
                message: Unauthorized
    Forbidden:
      description: Forbidden or insufficient permissions
      content:
        application/json:
          schema:
            allOf:
              - $ref: "#/components/schemas/ApiError"
              - type: object
                properties:
                  errors:
                    items:
                      properties:
                        code:
                          enum: [FORBIDDEN, INSUFFICIENT_PERMISSIONS]
          example:
            errors:
              - code: FORBIDDEN
                message: Forbidden
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            allOf:
              - $ref: "#/components/schemas/ApiError"
              - type: object
                properties:
                  errors:
                    items:
                      properties:
                        code:
                          enum: [NOT_FOUND]
          example:
            errors:
              - code: NOT_FOUND
                message: Resource not found
    TooManyRequests:
      description: Rate limit exceeded
      content:
        application/json:
          schema:
            allOf:
              - $ref: "#/components/schemas/ApiError"
              - type: object
                properties:
                  errors:
                    items:
                      properties:
                        code:
                          enum: [TOO_MANY_REQUESTS]
          example:
            errors:
              - code: TOO_MANY_REQUESTS
                message: Too many requests
...
    InternalServerError:
      description: Internal server error (e.g. failed to create/update/delete)
      content:
        application/json:
          schema:
            allOf:
              - $ref: "#/components/schemas/ApiError"
              - type: object
                properties:
                  errors:
                    items:
                      properties:
                        code:
                          enum: [INTERNAL_SERVER_ERROR, FAILED_TO_CREATE, FAILED_TO_UPDATE, FAILED_TO_DELETE, AUTH_FAILURE]
          example:
            errors:
              - code: INTERNAL_SERVER_ERROR
                message: Internal server error
  schemas:
    ApiError:
      type: object
      required: [errors]
      properties:
        errors:
          type: array
          minItems: 1
          items:
            type: object
            required: [code, message]
            properties:
              code:
                type: string
                enum:
                  - INTERNAL_SERVER_ERROR
                  - BAD_REQUEST
                  - NOT_FOUND
                  - UNAUTHORIZED
                  - FORBIDDEN
                  - FAILED_TO_DELETE
                  - INVALID_QUERY
                  - FAILED_TO_UPDATE
                  - FAILED_TO_CREATE
                  - AUTH_FAILURE
                  - VALIDATION_ERROR
                  - ALREADY_LOGGED_IN
                  - INSUFFICIENT_PERMISSIONS
                  - TOO_MANY_REQUESTS
                  - DUPLICATE_ENTRY
              message: { type: string }
              details:
                type: array
                description: Only present for VALIDATION_ERROR
                items:
                  type: object
                  properties:
                    field: { type: string }
                    validationMessage: { type: string, nullable: true }
...
    Operator:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        full_name: { type: string }
        parent_id: { type: integer, nullable: true }
        mnc: { type: integer, nullable: true }
...
    Location:
      type: object
      properties:
        id: { type: integer }
        city: { type: string, nullable: true }
        address: { type: string, nullable: true }
        longitude: { type: number }
        latitude: { type: number }
        updatedAt: { type: string, format: date-time }
        createdAt: { type: string, format: date-time }
...
    Station:
      type: object
      properties:
        id: { type: integer }
        station_id: { type: string }
        notes: { type: string, nullable: true }
        updatedAt: { type: string, format: date-time }
        createdAt: { type: string, format: date-time }
        is_confirmed: { type: boolean, nullable: true }
        operator: { $ref: "#/components/schemas/Operator", nullable: true }
        location: { $ref: "#/components/schemas/Location", nullable: true }
        extra_identificators: { $ref: "#/components/schemas/ExtraIdentificators", nullable: true }
...
    GsmCellDetails:
      type: object
      properties:
        lac: { type: integer }
        cid: { type: integer }
        e_gsm: { type: boolean, nullable: true }
    UmtsCellDetails:
      type: object
      properties:
        lac: { type: integer, nullable: true }
        arfcn: { type: integer, nullable: true }
        rnc: { type: integer }
        cid: { type: integer }
        cid_long: { type: integer }
    LteCellDetails:
      type: object
      properties:
        tac: { type: integer, nullable: true }
        enbid: { type: integer }
        clid: { type: integer }
        ecid: { type: integer }
        supports_nb_iot: { type: boolean, nullable: true }
    NrCellDetails:
      type: object
      properties:
        nrtac: { type: integer, nullable: true }
        gnbid: { type: integer, nullable: true }
        gnbid_length: { type: integer, nullable: true }
        clid: { type: integer, nullable: true }
        nci: { type: integer, nullable: true }
        pci: { type: integer, nullable: true }
        supports_nr_redcap: { type: boolean, nullable: true }
    UmtsCellDetailsInsert:
      description: UMTS cell details for create/update. Do not send cid_long; it is computed server-side from rnc and cid.
      type: object
      properties:
        lac: { type: integer, nullable: true }
        arfcn: { type: integer, nullable: true }
        rnc: { type: integer }
        cid: { type: integer }
    LteCellDetailsInsert:
      description: LTE cell details for create/update. Do not send ecid; it is computed server-side from enbid and clid.
      type: object
      properties:
        tac: { type: integer, nullable: true }
        enbid: { type: integer }
        clid: { type: integer }
        supports_nb_iot: { type: boolean, nullable: true }
    NrCellDetailsInsert:
      description: NR cell details for create/update. Do not send nci; it is computed server-side from gnbid, gnbid_length and clid.
      type: object
      properties:
        nrtac: { type: integer, nullable: true }
        gnbid: { type: integer, nullable: true }
        gnbid_length: { type: integer, nullable: true }
        clid: { type: integer, nullable: true }
        pci: { type: integer, nullable: true }
        supports_nr_redcap: { type: boolean, nullable: true }
    CellDetailsInsert:
      description: Cell RAT-specific payload for create/update. Omit server-calculated fields (cid_long for UMTS, ecid for LTE, nci for NR).
      oneOf:
        - $ref: "#/components/schemas/GsmCellDetails"
        - $ref: "#/components/schemas/UmtsCellDetailsInsert"
        - $ref: "#/components/schemas/LteCellDetailsInsert"
        - $ref: "#/components/schemas/NrCellDetailsInsert"
    CellDetails:
      oneOf:
        - $ref: "#/components/schemas/GsmCellDetails"
        - $ref: "#/components/schemas/UmtsCellDetails"
        - $ref: "#/components/schemas/LteCellDetails"
        - $ref: "#/components/schemas/NrCellDetails"
...
    Submission:
      type: object
      properties:
        id: { type: string, format: uuid }
        station_id: { type: integer, nullable: true }
        submitter_id: { type: string, format: uuid }
        status: { type: string, enum: [pending, approved, rejected] }
        type: { type: string, enum: [new, update, delete] }
        reviewer_id: { type: string, format: uuid, nullable: true }
        review_notes: { type: string, nullable: true }
        submitter_note: { type: string, nullable: true }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        reviewed_at: { type: string, format: date-time, nullable: true }
    SubmissionCreateSingle:
      description: In `cells[].details`, do not send cid_long (UMTS), ecid (LTE) or nci (NR); they are computed on the server.
      type: object
      properties:
        station_id: { type: integer, nullable: true }
        type: { type: string, enum: [new, update, delete] }
        submitter_note: { type: string, nullable: true }
        station: { $ref: "#/components/schemas/ProposedStationInsert", nullable: true }
        location: { $ref: "#/components/schemas/ProposedLocationInsert", nullable: true }
        cells: { type: array, items: { $ref: "#/components/schemas/ProposedCellInsert" }, nullable: true }
    ProposedStationInsert:
      type: object
      properties:
        operation: { type: string, enum: [add, update, delete], default: add }
        target_station_id: { type: integer, nullable: true }
        station_id: { type: string, nullable: true }
        operator_id: { type: integer, nullable: true }
        notes: { type: string, nullable: true }
        networks_id: { type: integer, nullable: true }
        networks_name: { type: string, nullable: true, maxLength: 150 }
        mno_name: { type: string, nullable: true, maxLength: 50 }
    ProposedLocationInsert:
      type: object
      required: [region_id, longitude, latitude]
      properties:
        region_id: { type: integer }
        city: { type: string, nullable: true }
        address: { type: string, nullable: true }
        longitude: { type: number }
        latitude: { type: number }
    ProposedCellInsert:
      description: Do not send cid_long (UMTS), ecid (LTE) or nci (NR) in details; they are computed on the server.
      type: object
      properties:
        target_cell_id: { type: integer, nullable: true }
        station_id: { type: integer, nullable: true }
        band_id: { type: integer, nullable: true }
        rat: { type: string, enum: [GSM, UMTS, LTE, NR], nullable: true }
        notes: { type: string, nullable: true }
        operation: { type: string, enum: [added, updated, removed] }
        details: { $ref: "#/components/schemas/CellDetailsInsert", nullable: true }
    SubmissionWithRelations:
      allOf:
        - $ref: "#/components/schemas/Submission"
        - type: object
          properties:
            station: { $ref: "#/components/schemas/Station", nullable: true }
            submitter: { $ref: "#/components/schemas/UserPublic" }
            reviewer: { $ref: "#/components/schemas/UserPublic", nullable: true }
            cells: { type: array, items: { $ref: "#/components/schemas/ProposedCellWithDetails" } }
            proposedStation: { $ref: "#/components/schemas/ProposedStation", nullable: true }
...
    SubmissionCreateResponse:
      allOf:
        - $ref: "#/components/schemas/Submission"
        - type: object
          properties:
            proposedStation: { $ref: "#/components/schemas/ProposedStation", nullable: true }
            proposedLocation: { $ref: "#/components/schemas/ProposedLocation", nullable: true }
            cells: { type: array, items: { $ref: "#/components/schemas/ProposedCellWithDetails" }, nullable: true }
            pending_photos: { type: integer, nullable: true, description: "Number of photos expected to be uploaded after submission creation" }
...
    ProposedCellWithDetails:
      type: object
      properties:
        id: { type: integer }
        submission_id: { type: string, format: uuid }
        operation: { type: string, enum: [add, update, delete] }
        target_cell_id: { type: integer, nullable: true }
        station_id: { type: integer, nullable: true }
        band_id: { type: integer, nullable: true }
        rat: { type: string, nullable: true }
        notes: { type: string, nullable: true }
        is_confirmed: { type: boolean, nullable: true }
        details: { $ref: "#/components/schemas/CellDetails", nullable: true }
...
    ProposedStation:
      type: object
      properties:
        id: { type: integer }
        submission_id: { type: string, format: uuid, nullable: true }
        operation: { type: string, enum: [add, update, delete] }
        target_station_id: { type: integer, nullable: true }
        station_id: { type: string, nullable: true }
        operator_id: { type: integer, nullable: true }
        notes: { type: string, nullable: true }
        networks_id: { type: integer, nullable: true }
        networks_name: { type: string, nullable: true }
        mno_name: { type: string, nullable: true }
        is_confirmed: { type: boolean, nullable: true }
        status: { type: string, enum: [published, inactive, pending] }
        updatedAt: { type: string, format: date-time }
        createdAt: { type: string, format: date-time }
    ProposedLocation:
      type: object
      properties:
        id: { type: integer }
        submission_id: { type: string, format: uuid, nullable: true }
        region_id: { type: integer }
        city: { type: string, nullable: true }
        address: { type: string, nullable: true }
        longitude: { type: number }
        latitude: { type: number }
        updatedAt: { type: string, format: date-time }
        createdAt: { type: string, format: date-time }
    UserPublic:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string, nullable: true }
        image: { type: string, nullable: true }
        displayUsername: { type: string, nullable: true }
...
    ExtraIdentificators:
      type: object
      properties:
        id: { type: integer }
        networks_id: { type: integer, nullable: true }
        networks_name: { type: string, nullable: true }
        mno_name: { type: string, nullable: true }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...